feat: refactor to v2
BREAKING CHANGE: this version requires that workflows that use this
v2 action use `actions/checkout` or similar prior to invoking this
action.  This is the standard way that most actions work, and avoids the
previously awkward fake checkout that this action had to do.

In the process, we also refactor the action and its CI tests.

The entire logic is no longer in single complex bash script that has to
worry about output formatting. Instead each stage is properly expressed
as an action step, which is a) simpler and b) much more similar with
workflows we already write.

The CI tests now execute the embedded action directly, rather than just
running the bash script. This is full intergration test of the syntax
and correctness of action.yml, include the creation of issues when
`opensafely check` fails.

Also, switch to using the gh cli for creating issues, rather than curl,
and spruced up the issue message a bit.